Fix colseg length check overflow on 32-bit targets
len(cells) > math.MaxUint32 compiles on 64-bit, where int is 64 bits,
but on a 32-bit target int is 32 bits and the untyped constant overflows
it, so the cross-compiled 386 and arm release builds failed. Widen the
length to uint64 for the comparison, which the bound is about anyway: the
count has to fit the uint32 length field on disk.

+++ colseg/colseg.go
@@ -78,7 +78,7 @@ const (
 // cell whose value does not fit that plane, or a composite type, forces the union
 // plane so the result is always correct regardless of the declared type.
 func Encode(valueType value.Type, cells []Cell) ([]byte, error) {
-	if len(cells) > math.MaxUint32 {
+	if uint64(len(cells)) > math.MaxUint32 {
 		return nil, ErrBadSegment
 	}
 	present := make([]value.Value, 0, len(cells))
